move shared fixtures to conftest
